requirements.txt
urllib3==1.22
Flask==0.12.2
flask-cors==3.0.3
flask_restful
flask_script==2.0.6
pyyaml==3.12
flasgger==0.8.1
flask-marshmallow==0.8.0
requests==2.18.4
Flask==0.12.2
humanize==0.5.1
swagger_marshmallow_codegen==0.4.0
gunicorn==19.7.1
werkzeug

dill==0.2.9
numpy==1.16.2
pandas==0.24.1
scikit-learn==0.20.3
xgboost==0.90
keras==2.2.4
lime==0.1.1.34